Binary now called rescript-editor-support.exe.
